feat: move arg-related functions to args
